[SECURITY] Fix possible XSS attack
The request to the extracts api is made with plaintext to get an
unformatted output (without bold, italic and other text formatting).
However, the output isn't parsed or anything else, it is passed directly
to the output as plain HTML. This allows an attacker to insert scripts
and other html into the output which is escaped on the page directly,
but not in the quicksearchlookup information box at the right side, which
is a security risk.

This change fixes the risk by not passing the content to an OOUI\HtmlSnippet
element anymore.

Bug: T130468
Change-Id: I430e137217c0f01b6807925e703ca71ce9989fe5