Mock Api request results based on other extensions
The extensions arebn't loaded during tests, and the test should not
expect, that these extensions are installed. The result of the required
method is now mocked. However, because it was private before, it is now
protected to make it testable. Not the best solution, but it's ok for this
method and the easiest solution here.

Bug: T163631
Change-Id: I128719efe1b67d36d72a23d47b40188a1e62440a